VSCODE-NLP-024 View Rule Fired, Created classes for sequence and logs
